feat(headless): remove renderElement, consolidate into useRender
renderElement has no callers left after the primitive migration. Move the
shared prop types (RenderProp, ComponentProps, DefaultProps,
StateAttributesMapping) and mergeProps into use-render.tsx, repoint the utils
barrel there, and delete render-element.tsx. The mergeProps unit tests and the
ComponentProps type test are ported alongside.
