+++ packages/headless/src/utils/index.ts
@@ -1,4 +1,3 @@
 export { cssVars } from './css-vars';
 export { resetLayoutStyles } from './reset-layout-styles';
-export { type ComponentProps, type DefaultProps, mergeProps, type RenderProp, renderElement } from './render-element';
-export { useRender } from './use-render';
+export { type ComponentProps, type DefaultProps, mergeProps, type RenderProp, useRender } from './use-render';
